pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "claude-eval"
version = "0.1.0"
description = "A lightweight evaluation harness for LLM outputs, built on the Claude API."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
claude-eval = "claude_eval.cli:main"

[tool.setuptools.packages.find]
include = ["claude_eval*"]
